# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BitBox API",
    "description": "Rust API for 360 Project notation online",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/config/user": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "summary": "Get current user config",
        "description": "This endpoint returns the current user (teacher) configuration.",
        "operationId": "get_config_by_user_id_route",
        "responses": {
          "200": {
            "description": "The current user configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserConfig"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "examples": {
                  "InvalidToken": {
                    "value": "Error"
                  },
                  "NoToken": {
                    "value": "Token not provided"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                },
                "example": "Database record"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Configuration"
        ],
        "summary": "Update current user configuration",
        "description": "This endpoint updates the configuration of the current user (teacher).",
        "operationId": "update_user_config_route",
        "requestBody": {
          "description": "The updated user configuration object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatedUserConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The user configuration updated successfully"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "examples": {
                  "InvalidToken": {
                    "value": "Error"
                  },
                  "NoToken": {
                    "value": "Token not provided"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                },
                "example": "Database record"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Alert": {
        "type": "object",
        "required": [
          "before_event",
          "hours"
        ],
        "properties": {
          "before_event": {
            "type": "boolean"
          },
          "hours": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "InternalError": {
        "default": null,
        "nullable": true
      },
...
      "NotFoundError": {
        "type": "object",
        "required": [
          "resource"
        ],
        "properties": {
          "resource": {
            "type": "string"
          }
        }
      },
...
      "UnauthorizedError": {
        "default": null,
        "nullable": true
      },
...
      "UpdatedUserConfig": {
        "type": "object",
        "properties": {
          "alerts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Alert"
            },
            "nullable": true
          }
        }
      },
      "UserConfig": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "alerts",
          "updated_at"
        ],
        "properties": {
          "alerts": {
            "type": "array",
            "items": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Alert"
                }
              ],
              "nullable": true
            }
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "api_key": [
        "read",
        "write"
      ]
    }
  ]
}